+++ rotkehlchen/tests/utils/ethereum.py
@@ -27,8 +27,19 @@ INFURA_TEST = random.choice([
     'https://mainnet.infura.io/v3/edeb337c7f41425e933ec619f3c5b940',
     'https://mainnet.infura.io/v3/66302b8fb9874614905a3cbe903a0dbb',
 ])
+ALCHEMY_TEST = 'https://eth-mainnet.alchemyapi.io/v2/ga1GtB7R26UgzjextaVpbaWZ49nSi2zt'
 
 ETHERSCAN_AND_INFURA_PARAMS: Tuple[str, List[Tuple]] = ('ethrpc_endpoint,ethereum_manager_connect_at_start, call_order', [  # noqa: E501
+    ('', (), (NodeName.ETHERSCAN,)),
+    (
+        INFURA_TEST,
+        (NodeName.OWN,),
+        (NodeName.OWN,),
+    ),
+])
+
+
+ETHERSCAN_AND_INFURA_AND_ALCHEMY: Tuple[str, List[Tuple]] = ('ethrpc_endpoint,ethereum_manager_connect_at_start, call_order', [  # noqa: E501
     # Query etherscan only
     ('', (), (NodeName.ETHERSCAN,)),
     # For "our own" node querying use infura
@@ -37,8 +48,15 @@ ETHERSCAN_AND_INFURA_PARAMS: Tuple[str, List[Tuple]] = ('ethrpc_endpoint,ethereu
         (NodeName.OWN,),
         (NodeName.OWN,),
     ),
+    (
+        ALCHEMY_TEST,
+        (NodeName.OWN,),
+        (NodeName.OWN,),
+    ),
 ])
 
+
+# Test with etherscan and infura
 ETHEREUM_TEST_PARAMETERS: Tuple[str, List[Tuple]]
 if 'GITHUB_WORKFLOW' in os.environ:
     # For Github actions don't use infura. It seems that connecting to it
@@ -50,6 +68,20 @@ if 'GITHUB_WORKFLOW' in os.environ:
 else:
     # For Travis and local tests also use Infura, works fine
     ETHEREUM_TEST_PARAMETERS = ETHERSCAN_AND_INFURA_PARAMS
+
+
+# Test with multipe node types and etherscan
+ETHEREUM_FULL_TEST_PARAMETERS: Tuple[str, List[Tuple]]
+if 'GITHUB_WORKFLOW' in os.environ:
+    # For Github actions don't use infura. It seems that connecting to it
+    # from Github actions hangs and times out
+    ETHEREUM_FULL_TEST_PARAMETERS = ('ethrpc_endpoint,ethereum_manager_connect_at_start, call_order', [  # noqa: E501
+        # Query etherscan only
+        ('', (), (NodeName.ETHERSCAN,)),
+    ])
+else:
+    # For Travis and local tests also use Infura, works fine
+    ETHEREUM_FULL_TEST_PARAMETERS = ETHERSCAN_AND_INFURA_AND_ALCHEMY
 
 
 def wait_until_all_nodes_connected(