+++ docs/changelog.rst
@@ -18,6 +18,7 @@ Changelog
 * :bug:`3667` Users should be able to upload asset icons to docker instances when using a remote connection via the application.
 * :bug:`3664` Binance US users will no longer see errors regarding the fiat payments and orders endpoints.
 * :bug:`3666` An ethereum token's address will be properly filled when navigating to the asset edit from the asset overview page.
+* :bug:`-` Transaction receipts will now be processed properly even for nodes(such as open ethereum) that do not return the type field for non EIP1559 transactions.
 
 * :release:`1.21.3 <2021-10-28>`
 * :bug:`2178` Premium DB sync popup should no longer popup if you only use one instance of rotki in one system.